+++ src/ewokscore/_serialization/yaml.py
@@ -1,38 +1,66 @@
 from typing import Any
-from typing import List
+from typing import Callable
+from typing import Dict
 from typing import Optional
 from typing import TextIO
+from typing import Union
 
 import yaml
 
 from . import common
+from .common import types
 
 
 def dumps(
-    obj: Any, custom_rules: Optional[List[common.RuleType]] = None, **kwargs
+    obj: Any,
+    item_serializers: Optional[Dict[types.Path, types.ConverterType]] = None,
+    serializer: Optional[Union[types.GraphSerializer, str]] = None,
+    insert_serialize_info: Optional[
+        Callable[[Any, str, types.SerializeInfo], Any]
+    ] = None,
+    **kwargs,
 ) -> str:
     """
     Serialize Python object to YAML string.
     """
-    pre = common.pre_serialize(obj, custom_rules=custom_rules)
-    return yaml.dump(pre, stream=None, Dumper=yaml.Dumper, **kwargs)
+    pre = common.pre_serialize(
+        obj,
+        item_serializers=item_serializers,
+        serializer=serializer,
+        insert_serialize_info=insert_serialize_info,
+    )
+    return yaml.dump(pre, stream=None, Dumper=yaml.SafeDumper, **kwargs)
 
 
 def dump(
     obj: Any,
     fp: TextIO,
-    custom_rules: Optional[List[common.RuleType]] = None,
+    item_serializers: Optional[Dict[types.Path, types.ConverterType]] = None,
+    serializer: Optional[Union[types.GraphSerializer, str]] = None,
+    insert_serialize_info: Optional[
+        Callable[[Any, str, types.SerializeInfo], Any]
+    ] = None,
     **kwargs,
 ) -> None:
     """
     Write Python object to YAML file-like object.
     """
-    pre = common.pre_serialize(obj, custom_rules=custom_rules)
-    yaml.dump(pre, stream=fp, Dumper=yaml.Dumper, **kwargs)
+    pre = common.pre_serialize(
+        obj,
+        item_serializers=item_serializers,
+        serializer=serializer,
+        insert_serialize_info=insert_serialize_info,
+    )
+    yaml.dump(pre, stream=fp, Dumper=yaml.SafeDumper, **kwargs)
 
 
 def loads(
-    s: str, custom_rules: Optional[List[common.RuleType]] = None, **kwargs
+    s: str,
+    item_deserializers: Optional[Dict[types.Path, types.ConverterType]] = None,
+    pop_serialize_info: Optional[
+        Callable[[Any, str], Optional[types.SerializeInfo]]
+    ] = None,
+    **kwargs,
 ) -> Any:
     """
     Deserialize YAML string to Python object.
@@ -40,13 +68,22 @@ def loads(
     try:
         raw = yaml.load(s, yaml.Loader, **kwargs)
     except (yaml.YAMLError, TypeError) as e:
-        raise common.EwoksDecodeError from e
+        raise types.EwoksDecodeError from e
 
-    return common.post_deserialize(raw, custom_rules=custom_rules)
+    return common.post_deserialize(
+        raw,
+        item_deserializers=item_deserializers,
+        pop_serialize_info=pop_serialize_info,
+    )
 
 
 def load(
-    fp: TextIO, custom_rules: Optional[List[common.RuleType]] = None, **kwargs
+    fp: TextIO,
+    item_deserializers: Optional[Dict[types.Path, types.ConverterType]] = None,
+    pop_serialize_info: Optional[
+        Callable[[Any, str], Optional[types.SerializeInfo]]
+    ] = None,
+    **kwargs,
 ) -> Any:
     """
     Load Python object from YAML file-like object.
@@ -54,6 +91,10 @@ def load(
     try:
         raw = yaml.load(fp, yaml.Loader, **kwargs)
     except (yaml.YAMLError, TypeError) as e:
-        raise common.EwoksDecodeError from e
+        raise types.EwoksDecodeError from e
 
-    return common.post_deserialize(raw, custom_rules=custom_rules)
+    return common.post_deserialize(
+        raw,
+        item_deserializers=item_deserializers,
+        pop_serialize_info=pop_serialize_info,
+    )