+++ src/ewokscore/events/handlers/__init__.py
@@ -1,5 +1,6 @@
 # fmt: off
-from ewoksutils.import_utils import instantiate_class as instantiate_handler  # noqa F401
+from ewoksutils.import_utils import \
+    instantiate_class as instantiate_handler  # noqa F401
 
 from .base import *  # noqa F401
 from .sqlite3 import Sqlite3EwoksEventHandler  # noqa F401