+++ src/ewokscore/tests/serialization/example_data/types.py
@@ -1,0 +1,6 @@
+class CustomType:
+    def __init__(self, value):
+        self._value = value
+
+    def __eq__(self, value):
+        return isinstance(value, CustomType) and value._value == self._value