+++ src/ewokscore/tests/serialization/test_networkx.py
@@ -5,6 +5,7 @@ import yaml
 
 from ...graph import serialize
 from . import example_data
+from .example_data import compare_json
 
 
 def test_yaml_graph_serialization(tmp_path):
@@ -13,14 +14,14 @@ def test_yaml_graph_serialization(tmp_path):
     G.add_node("node1", default_args=original_data)
 
     dest = tmp_path / "workflow.yaml"
-    serialize.dump(G, destination=dest, representation="yaml")
+    serialize.dump(G, destination=dest, representation="yaml", serializer="json_pickle")
 
     python_data = serialize.load(dest).nodes["node1"]["default_args"]
-    example_data.assert_python_data(python_data, original_data)
+    compare_json.assert_deserialized_data(python_data, original_data)
 
     raw_data = yaml.safe_load(dest.read_text(encoding="utf-8"))
     raw_data = raw_data["nodes"][0]["default_args"]
-    example_data.assert_raw_data(raw_data, original_data)
+    compare_json.assert_serialized_data(raw_data, original_data)
 
 
 def test_json_graph_serialization(tmp_path):
@@ -29,11 +30,11 @@ def test_json_graph_serialization(tmp_path):
     G.add_node("node1", default_args=original_data)
 
     dest = tmp_path / "workflow.json"
-    serialize.dump(G, destination=dest, representation="json")
+    serialize.dump(G, destination=dest, representation="json", serializer="json_pickle")
 
     python_data = serialize.load(dest).nodes["node1"]["default_args"]
-    example_data.assert_python_data(python_data, original_data)
+    compare_json.assert_deserialized_data(python_data, original_data)
 
     raw_data = json.loads(dest.read_text(encoding="utf-8"))
     raw_data = raw_data["nodes"][0]["default_args"]
-    example_data.assert_raw_data(raw_data, original_data)
+    compare_json.assert_serialized_data(raw_data, original_data)